+++ packages/frameworks/preact/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @zag-js/preact
 
+## 1.12.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/core@1.12.2
+  - @zag-js/store@1.12.2
+  - @zag-js/types@1.12.2
+
+## 1.12.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/core@1.12.1
+  - @zag-js/store@1.12.1
+  - @zag-js/types@1.12.1
+
 ## 1.12.0
 
 ### Patch Changes