+++ packages/machines/floating-panel/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @zag-js/floating-panel
 
+## 1.12.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/anatomy@1.12.2
+  - @zag-js/core@1.12.2
+  - @zag-js/store@1.12.2
+  - @zag-js/types@1.12.2
+  - @zag-js/utils@1.12.2
+  - @zag-js/dismissable@1.12.2
+  - @zag-js/dom-query@1.12.2
+  - @zag-js/popper@1.12.2
+  - @zag-js/rect-utils@1.12.2
+
+## 1.12.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/anatomy@1.12.1
+  - @zag-js/core@1.12.1
+  - @zag-js/store@1.12.1
+  - @zag-js/types@1.12.1
+  - @zag-js/utils@1.12.1
+  - @zag-js/dismissable@1.12.1
+  - @zag-js/dom-query@1.12.1
+  - @zag-js/popper@1.12.1
+  - @zag-js/rect-utils@1.12.1
+
 ## 1.12.0
 
 ### Patch Changes