+++ packages/machines/tags-input/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @zag-js/tags-input
 
+## 1.12.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/anatomy@1.12.2
+  - @zag-js/core@1.12.2
+  - @zag-js/types@1.12.2
+  - @zag-js/auto-resize@1.12.2
+  - @zag-js/utils@1.12.2
+  - @zag-js/dom-query@1.12.2
+  - @zag-js/interact-outside@1.12.2
+  - @zag-js/live-region@1.12.2
+
+## 1.12.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/anatomy@1.12.1
+  - @zag-js/core@1.12.1
+  - @zag-js/types@1.12.1
+  - @zag-js/auto-resize@1.12.1
+  - @zag-js/utils@1.12.1
+  - @zag-js/dom-query@1.12.1
+  - @zag-js/interact-outside@1.12.1
+  - @zag-js/live-region@1.12.1
+
 ## 1.12.0
 
 ### Patch Changes