+++ packages/utilities/color-utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @zag-js/color-utils
 
+## 1.12.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/utils@1.12.2
+
+## 1.12.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/utils@1.12.1
+
 ## 1.12.0
 
 ### Patch Changes