+++ packages/utilities/dismissable/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @zag-js/dismissable
 
+## 1.12.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/utils@1.12.2
+  - @zag-js/dom-query@1.12.2
+  - @zag-js/interact-outside@1.12.2
+
+## 1.12.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/utils@1.12.1
+  - @zag-js/dom-query@1.12.1
+  - @zag-js/interact-outside@1.12.1
+
 ## 1.12.0
 
 ### Patch Changes