+++ packages/utilities/file-utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @zag-js/file-utils
 
+## 1.12.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/i18n-utils@1.12.2
+
+## 1.12.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/i18n-utils@1.12.1
+
 ## 1.12.0
 
 ### Patch Changes